requirements.txt
Pillow==11.1.0
torch==2.6.0
torchvision==0.21.0
torchvision==0.21.0
